split employees files into subclass files
