+++ src/employees.h
@@ -1,10 +1,6 @@
 #ifndef EMPLOYEES_H
 #define EMPLOYEES_H
 
-#include <string>
-#include <map>
-#include <memory>
-#include <iostream>
 #include "order.h"
 #include "material.h"
 #include "reports.h"
@@ -14,95 +10,28 @@
 #include "clocks.h"
 #include "employeemanager.h"
 
-class OrderManager;
-class IClock;
-class EmployeeManager;
-
-class Employee {
-    protected:
-        OrderManager* order_manager;
-
-    public:
-        int emp_id = 0;
-        std::string emp_name;
-
-        Employee(OrderManager* order_manager, std::string emp_name) : order_manager(order_manager), emp_name(emp_name) { }
-
-        virtual std::string getEmpType() = 0;
-
-        virtual ~Employee() = default;
-};
+#include <string>
+#include <map>
+#include <memory>
+#include <iostream>
 
 
 //Note: most likely, when there are no employees, the
 //first employee should always be the administrator
 
-class ReceptReportManager;
-class Receptionist: public Employee {
-    private:
-        IClock& clock;
-        ReceptReportManager* receptreport_manager;
+class Employee {
+    protected:
+        InMemoryOrderManager* order_manager;
 
     public:
-        Receptionist(OrderManager* order_manager, std::string emp_name, IClock& clock, ReceptReportManager* receptreport_manager) : Employee(order_manager, emp_name), clock(clock), receptreport_manager(receptreport_manager) { }
+        int emp_id = 0;
+        std::string emp_name;
 
-        std::string getEmpType();
+        Employee(InMemoryOrderManager* order_manager, std::string emp_name) : order_manager(order_manager), emp_name(emp_name) { }
 
-        int makeOrder(std::shared_ptr<Client> client, Service service, unsigned int in_x_days);
+        virtual std::string getEmpType() = 0;
 
-        void assignOrder(Order* assign, int emp_id);
-
-        int submitReport();
-};
-
-class MaterialManager;
-class PhotoReportManager;
-class Photographer: public Employee {
-    private:
-        IClock& clock;
-        MaterialManager* material_manager;
-        PhotoReportManager* photoreport_manager;
-        MaterialManager* consumed_materials;
-
-        std::map<std::shared_ptr<Material>, int> getConsumedMaterials();
-    public:
-        Photographer(OrderManager* order_manager, std::string emp_name, IClock& clock, MaterialManager* material_manager, PhotoReportManager* photoreport_manager) : Employee(order_manager, emp_name), clock(clock), material_manager(material_manager), photoreport_manager(photoreport_manager) {
-            this->consumed_materials = new MaterialManager();
- }
-
-        std::string getEmpType();
-
-        void switchOrderStatus(Order* changedorder, CompletionStatus compl_status);
-
-        void consumeMaterial(std::string mat_type, unsigned int quantity);
-
-
-        int submitReport();
-
-        ~Photographer() override;
-};
-
-
-class Administrator: public Employee {
-    private:
-        MaterialManager* material_manager;
-        ReceptReportManager* receptreport_manager;
-        PhotoReportManager* photoreport_manager;
-
-    public:
-        Administrator(OrderManager* order_manager, std::string emp_name, MaterialManager* material_manager, ReceptReportManager* receptreport_manager, PhotoReportManager* photoreport_manager) : Employee(order_manager, emp_name), material_manager(material_manager), receptreport_manager(receptreport_manager), photoreport_manager(photoreport_manager) { }
-
-        std::string getEmpType();
-
-        std::map<int, std::shared_ptr<ReceptReport>> listReceptReports();
-
-        std::map<int, std::shared_ptr<PhotoReport>> listPhotoReports();
-
-        std::vector<std::shared_ptr<Material>> listMaterials();
-
-        void addMaterial(std::string mat_type, unsigned int quantity);
-
-        void removeMaterial(std::string mat_type);
+        virtual ~Employee() = default;
 };
 
 #endif
